Move inspection table to new file
