+++ frontend/src/components/Pages/InspectionPage/InspectionSection.tsx
@@ -1,4 +1,4 @@
-import { Table, Card, Typography, Icon, Button, EdsProvider } from '@equinor/eds-core-react'
+import { Card, Typography, Icon, Button, EdsProvider } from '@equinor/eds-core-react'
 import styled from 'styled-components'
 import { useLanguageContext } from 'components/Contexts/LanguageContext'
 import { useState, useEffect } from 'react'
@@ -7,15 +7,15 @@ import { Deck } from 'models/Deck'
 import { useInstallationContext } from 'components/Contexts/InstallationContext'
 import { RefreshProps } from '../FrontPage/FrontPage'
 import { tokens } from '@equinor/eds-tokens'
-import { CondensedMissionDefinition, MissionDefinition } from 'models/MissionDefinition'
+import { CondensedMissionDefinition } from 'models/MissionDefinition'
 import { useNavigate } from 'react-router-dom'
-import { config } from 'config'
 import { ScheduleMissionDialog } from './ScheduleMissionDialog'
-import { Icons } from 'utils/icons'
 import { MissionStatus } from 'models/Mission'
+import { getInspectionDeadline } from 'utils/StringFormatting'
+import { InspectionTable } from './InspectionTable'
 
 const StyledCard = styled(Card)`
-    width: 250px;
+    width: 280px;
     padding: 12px;
     border-radius: 20px;
     :hover {
@@ -37,60 +37,18 @@ const StyledDeckCards = styled.div`
     gap: 1rem;
 `
 
-const TableWithHeader = styled.div`
-    gap: 2rem;
-`
-
 const StyledContent = styled.div`
     display: flex;
     flex-direction: column;
     gap: 1rem;
 `
 
-const StyledIcon = styled(Icon)`
-    display: flex;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-    scale: 50%;
-`
-
-const Circle = (fill: string) => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="13" height="14" viewBox="0 0 13 14" fill="none">
-        <circle cx="6.5" cy="7" r="6.5" fill={fill} />
-    </svg>
-)
-const RedCircle = Circle('#EB0000')
-const YellowCircle = Circle('#FF9200')
-const GreenCircle = Circle('#4BB748')
-
-interface DeckMissionType {
+export interface DeckMissionType {
     [deckId: string]: { missionDefinitions: CondensedMissionDefinition[]; deck: Deck }
 }
 
-interface OngoingMissionType {
+export interface OngoingMissionType {
     [missionId: string]: boolean
-}
-
-const formatBackendDateTimeToDate = (date: Date) => {
-    return new Date(date.toString())
-}
-
-const getInspectionDeadline = (inspectionFrequency: string, lastRunTime: Date): Date => {
-    const dayHourSecondsArray = inspectionFrequency.split(':')
-    const days: number = +dayHourSecondsArray[0]
-    const hours: number = +dayHourSecondsArray[1]
-    const minutes: number = +dayHourSecondsArray[2]
-
-    lastRunTime = formatBackendDateTimeToDate(lastRunTime)
-
-    let deadline = lastRunTime
-    deadline.setDate(deadline.getDate() + days)
-    deadline.setHours(deadline.getHours() + hours)
-    deadline.setMinutes(deadline.getMinutes() + minutes)
-    return deadline
-    // More flexibly we can also define the deadline in terms of milliseconds:
-    // new Date(lastRunTime.getTime() + (1000 * 60 * days) + (1000 * 60 * 60 * hours) + (1000 * 60 * 60 * 24 * days))
 }
 
 export function InspectionSection({ refreshInterval }: RefreshProps) {
@@ -153,137 +111,6 @@ export function InspectionSection({ refreshInterval }: RefreshProps) {
         return () => clearInterval(id)
     }, [deckMissions])
 
-    const getInspectionStatus = (inspectionFrequency: string, lastRunTime: Date) => {
-        const deadlineDate = getInspectionDeadline(inspectionFrequency, lastRunTime)
-        // The magical number on the right is the number of milliseconds in a day
-        const deadline = new Date(deadlineDate.getTime() - new Date().getTime()).getTime() / 8.64e7
-
-        if (deadline <= 0) {
-            return (
-                <>
-                    {RedCircle} {TranslateText('Past deadline')}
-                </>
-            )
-        } else if (deadline > 0 && deadline <= 1) {
-            return (
-                <>
-                    {RedCircle} {TranslateText('Due today')}
-                </>
-            )
-        } else if (deadline > 1 && deadline <= 7) {
-            return (
-                <>
-                    {YellowCircle} {TranslateText('Due this week')}
-                </>
-            )
-        } else if (deadline > 7 && deadline <= 14) {
-            return (
-                <>
-                    {YellowCircle} {TranslateText('Due within two weeks')}
-                </>
-            )
-        } else if (deadline > 7 && deadline <= 30) {
-            return (
-                <>
-                    {GreenCircle} {TranslateText('Due within a month')}
-                </>
-            )
-        }
-        return (
-            <>
-                {GreenCircle} {TranslateText('Up to date')}
-            </>
-        )
-    }
-
-    const formatDateString = (dateStr: Date) => {
-        let newStr = dateStr.toString()
-        newStr = newStr.slice(0, newStr.length - 8)
-        newStr = newStr.replaceAll('T', ' ')
-        return newStr
-    }
-
-    const getInspectionRow = (mission: CondensedMissionDefinition) => {
-        let status
-        let lastCompleted: string = ''
-        let deadline: string = ''
-        const isScheduled = Object.keys(ongoingMissions).includes(mission.id) && ongoingMissions[mission.id]
-        if (isScheduled) {
-            status = (
-                <>
-                    {GreenCircle} {TranslateText('Already scheduled')}
-                </>
-            )
-        } else {
-            if (!mission.lastRun || !mission.lastRun.endTime) {
-                status = (
-                    <>
-                        {RedCircle} {TranslateText('Not yet performed')}
-                    </>
-                )
-                lastCompleted = TranslateText('Never')
-            } else if (mission.inspectionFrequency) {
-                status = getInspectionStatus(mission.inspectionFrequency, mission.lastRun.endTime!)
-                lastCompleted = formatDateString(mission.lastRun.endTime!)
-                deadline = getInspectionDeadline(mission.inspectionFrequency, mission.lastRun.endTime!).toDateString()
-            } else {
-                status = TranslateText('No planned inspection')
-                lastCompleted = formatDateString(mission.lastRun.endTime!)
-            }
-        }
-
-        return (
-            <Table.Row key={mission.id}>
-                <Table.Cell>{status}</Table.Cell>
-                <Table.Cell>
-                    <Typography
-                        link
-                        onClick={() => navigate(`${config.FRONTEND_BASE_ROUTE}/mission-definition/${mission.id}`)}
-                    >
-                        {mission.name}
-                    </Typography>
-                </Table.Cell>
-                <Table.Cell>{mission.comment}</Table.Cell>
-                <Table.Cell>{lastCompleted}</Table.Cell>
-                <Table.Cell>{deadline}</Table.Cell>
-                <Table.Cell>
-                    <StyledIcon
-                        color={`${tokens.colors.interactive.focus.hex}`}
-                        name={Icons.AddOutlined}
-                        size={16}
-                        title={TranslateText('Add to queue')}
-                        onClick={() => {
-                            setisDialogOpen(true)
-                            setSelectedMissions([mission])
-                        }}
-                    />
-                </Table.Cell>
-            </Table.Row>
-        )
-    }
-
-    const getInspectionsTable = (deck: Deck) => (
-        <TableWithHeader>
-            <Typography variant="h3">
-                {TranslateText('Inspection Missions') + ' ' + TranslateText('for') + ' ' + deck.deckName}
-            </Typography>
-            <Table>
-                <Table.Head sticky>
-                    <Table.Row>
-                        <Table.Cell>{TranslateText('Status')}</Table.Cell>
-                        <Table.Cell>{TranslateText('Name')}</Table.Cell>
-                        <Table.Cell>{TranslateText('Description')}</Table.Cell>
-                        <Table.Cell>{TranslateText('Last completed')}</Table.Cell>
-                        <Table.Cell>{TranslateText('Deadline')}</Table.Cell>
-                        <Table.Cell>{TranslateText('Add to queue')}</Table.Cell>
-                    </Table.Row>
-                </Table.Head>
-                <Table.Body>
-                    {deckMissions[deck.id].missionDefinitions.map((mission) => getInspectionRow(mission))}
-                </Table.Body>
-            </Table>
-        </TableWithHeader>
-    )
 
     return (
         <>
@@ -315,7 +142,14 @@ export function InspectionSection({ refreshInterval }: RefreshProps) {
                         <Typography variant="h1">{TranslateText('No Deck Inspections Available')}</Typography>
                     )}
                 </StyledDeckCards>
-                {selectedDeck && getInspectionsTable(selectedDeck)}
+                {selectedDeck && 
+                    <InspectionTable 
+                        deck={selectedDeck} 
+                        openDialog={() => setisDialogOpen(true)}
+                        setSelectedMissions={setSelectedMissions}
+                        deckMissions={deckMissions}
+                        ongoingMissions={ongoingMissions} />
+                }
             </StyledContent>
             {isDialogOpen && (
                 <ScheduleMissionDialog
