+++ frontend/src/language/en.json
@@ -151,5 +151,6 @@
     "No inspection frequency set": "No inspection frequency set",
     "Are you sure you want to delete": "Are you sure you want to delete",
     "Yes": "Yes",
-    "Missions": "Missions"
+    "Missions": "Missions",
+    "Schedule all": "Schedule all"
 }
