+++ frontend/src/language/no.json
@@ -153,5 +153,6 @@
     "No inspection frequency set": "Ingen inspeksjonsfrekvens satt",
     "Are you sure you want to delete": "Er du sikker at du ønsker å slette",
     "Yes": "Ja",
-    "Missions": "Oppdrag"
+    "Missions": "Oppdrag",
+    "Schedule all": "Legg til alle i køen"
 }
